requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
APScheduler==3.9.1
apscheduler-di==0.0.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.0.6
environs==9.5.0
frozenlist==1.3.0
idna==3.3
marshmallow==3.17.0
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.2.0
psycopg-binary==3.0.15
psycopg2-binary==2.9.3
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
rodi==1.1.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-requests==2.28.0
types-urllib3==1.26.16
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
wrapt==1.14.1
yarl==1.7.2
